+++ bricks/e2e/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.47](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.46...@next-bricks/e2e@0.4.47) (2024-09-24)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.46](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.45...@next-bricks/e2e@0.4.46) (2024-09-23)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.45](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.44...@next-bricks/e2e@0.4.45) (2024-09-13)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.44](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.43...@next-bricks/e2e@0.4.44) (2024-09-06)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.43](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.42...@next-bricks/e2e@0.4.43) (2024-09-03)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.42](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.41...@next-bricks/e2e@0.4.42) (2024-09-02)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.41](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.40...@next-bricks/e2e@0.4.41) (2024-09-02)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
+## [0.4.40](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.39...@next-bricks/e2e@0.4.40) (2024-08-27)
+
+**Note:** Version bump only for package @next-bricks/e2e
+
+
+
+
+
 ## [0.4.39](https://github.com/easyops-cn/next-core/compare/@next-bricks/e2e@0.4.38...@next-bricks/e2e@0.4.39) (2024-08-19)
 
 **Note:** Version bump only for package @next-bricks/e2e