+++ packages/doc-helpers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14](https://github.com/easyops-cn/next-core/compare/@next-core/doc-helpers@0.2.13...@next-core/doc-helpers@0.2.14) (2024-08-27)
+
+**Note:** Version bump only for package @next-core/doc-helpers
+
+
+
+
+
 ## [0.2.13](https://github.com/easyops-cn/next-core/compare/@next-core/doc-helpers@0.2.12...@next-core/doc-helpers@0.2.13) (2024-06-21)
 
 **Note:** Version bump only for package @next-core/doc-helpers