+++ packages/monaco-contributions/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11](https://github.com/easyops-cn/next-core/compare/@next-core/monaco-contributions@0.3.10...@next-core/monaco-contributions@0.3.11) (2024-08-27)
+
+
+### Bug Fixes
+
+* fallback monaco-editor to v0.50.0 ([cb6f63f](https://github.com/easyops-cn/next-core/commit/cb6f63fb0c82721d518917c800b0a3a32ccbde2d))
+
+
+
+
+
+## [0.3.10](https://github.com/easyops-cn/next-core/compare/@next-core/monaco-contributions@0.3.9...@next-core/monaco-contributions@0.3.10) (2024-08-27)
+
+
+### Bug Fixes
+
+* downgrade monaco-editor to v0.50.0 ([8958bb1](https://github.com/easyops-cn/next-core/commit/8958bb1abcd479c48adcbdd9d220ab5c03581ae8))
+* upgrade monaco-editor v0.51.0 and make it compatible ([1a3da4d](https://github.com/easyops-cn/next-core/commit/1a3da4d065d10916e2be5689986aac3a497d9019))
+
+
+
+
+
 ## [0.3.9](https://github.com/easyops-cn/next-core/compare/@next-core/monaco-contributions@0.3.8...@next-core/monaco-contributions@0.3.9) (2024-06-21)
 
 **Note:** Version bump only for package @next-core/monaco-contributions