+++ packages/react-runtime/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.43](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.42...@next-core/react-runtime@1.6.43) (2024-09-24)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.42](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.41...@next-core/react-runtime@1.6.42) (2024-09-23)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.41](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.40...@next-core/react-runtime@1.6.41) (2024-09-13)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.40](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.39...@next-core/react-runtime@1.6.40) (2024-09-06)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.39](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.38...@next-core/react-runtime@1.6.39) (2024-09-02)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.38](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.37...@next-core/react-runtime@1.6.38) (2024-09-02)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
+## [1.6.37](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.36...@next-core/react-runtime@1.6.37) (2024-08-27)
+
+**Note:** Version bump only for package @next-core/react-runtime
+
+
+
+
+
 ## [1.6.36](https://github.com/easyops-cn/next-core/compare/@next-core/react-runtime@1.6.35...@next-core/react-runtime@1.6.36) (2024-08-19)
 
 **Note:** Version bump only for package @next-core/react-runtime